Journal Clubs
Journal clubs are now listed and searchable. They are defined by adding a markdown file to /_journal-clubs/ and can carry a host of information, including a description of the JC.

Journal clubs can be linked to explicitly by using #JCName in the URL.

IFF we feel it's worth it, we can make the people bits much fancier and allow links to people's individual websites/osf repos/githubs/etc etc.

Added example JC files for Oxford and Cambridge.