Rollcall should now work
Rollcall will attempt to update the oldest journal club each time it is called. When a journal club has been rollcalled, if it is alerted that updates are needed, it drops to the bottom of the priority list, and will not be re-called again until a month or so after being notified that updates are needed.
Each journal club gets three warnings before being marked as inactive.