Rollcall (#46)
Rollcall will attempt to update the oldest journal club each time it is called. When a journal club has been rollcalled, if it is alerted that updates are needed, it drops to the bottom of the priority list, and will not be re-called again until a month or so after being notified that updates are needed.
Each journal club gets three warnings before being marked as inactive.

To aid this process, additional metadata has been added to the journal-club.md files, a last-message-time and a last-message-level. Message levels start at 0 (up to date, to which they are reset when editing) and increase by 1 each time a notification is sent.