Updated organisers list and updated sign up sheet
We've now officially handed over to Sumbul and Bérengère and they have updated the sign-up sheet. I've also changed the address from the division of Psychiatry (where Niamh and I are based) to the more general University of Edinburgh location. The new organisers do not have a physical site at the moment.